revert(knockdog): #9918e1 원복
