+++ apps/knockdog/app/(auth)/signup/profile/pet-info/relation/page.tsx
@@ -1,20 +1,31 @@
 'use client';
 
 import { useEffect, useState } from 'react';
+import {
+  ActionButton,
+  Icon,
+  TextField,
+  TextFieldInput,
+  BottomSheet,
+} from '@knockdog/ui';
 import Link from 'next/link';
-import dynamic from 'next/dynamic';
-import { ActionButton, Icon, TextField, TextFieldInput } from '@knockdog/ui';
+import { cn } from '@knockdog/ui/lib';
 import { useHeaderContext } from '@widgets/Header';
 
-const ActionSheet = dynamic(() => import('./ActionSheet'), {
-  ssr: false,
-});
+const RELATION_OPTIONS = [
+  { label: '엄마', value: 'mother' },
+  { label: '아빠', value: 'father' },
+  { label: '가족', value: 'family' },
+  { label: '보호자', value: 'guardian' },
+  { label: '기타', value: 'etc' },
+];
 
 export default function PetInfoRelationPage() {
   const { setTitle } = useHeaderContext();
 
   const [petName, setPetName] = useState('살구');
 
+  const [relation, setRelation] = useState<string>('mother');
   const [customRelation, setCustomRelation] = useState<string>('');
   const [isOpen, setIsOpen] = useState(false);
 
@@ -74,7 +85,55 @@ export default function PetInfoRelationPage() {
         </div>
       </div>
       {/* bottom sheet container */}
-      <ActionSheet isOpen={isOpen} onOpenChange={setIsOpen} />
+      <BottomSheet.Root open={isOpen} onOpenChange={setIsOpen}>
+        <BottomSheet.Content>
+          <BottomSheet.Handle />
+          <BottomSheet.Header className='border-line-200 justify-between border-b'>
+            <BottomSheet.Title>관계 선택</BottomSheet.Title>
+            <BottomSheet.Close className='absolute right-4 flex items-center justify-center'>
+              <Icon icon='Close' />
+            </BottomSheet.Close>
+          </BottomSheet.Header>
+          <div className='px-6'>
+            {RELATION_OPTIONS.map((option) => (
+              <div
+                key={option.value}
+                className={cn(
+                  'body2-semibold flex cursor-pointer items-center justify-between py-4',
+                  relation === option.value && 'text-text-accent'
+                )}
+                onClick={() => setRelation(option.value)}
+              >
+                <label
+                  className='flex cursor-pointer items-center gap-x-2'
+                  htmlFor={option.value}
+                >
+                  {option.label}
+                </label>
+                <input
+                  type='radio'
+                  name='relation'
+                  value={option.value}
+                  checked={relation === option.value}
+                  onChange={() => setRelation(option.value)}
+                  id={option.value}
+                  className='sr-only'
+                />
+                <span
+                  className={cn(
+                    'relative m-[3px] inline-flex h-5 w-5 items-center justify-center rounded-full border-2 border-neutral-400 transition-colors',
+                    relation === option.value && 'border-text-accent'
+                  )}
+                >
+                  {relation === option.value && (
+                    <span className='bg-text-accent block h-3 w-3 rounded-full' />
+                  )}
+                </span>
+              </div>
+            ))}
+          </div>
+        </BottomSheet.Content>
+      </BottomSheet.Root>
     </>
   );
 }
